+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Changelog
 
+## v.0.7.0
+
+### Added
+
+- Ability to use Analytics provider's survey feature.
+
+## v.0.6.2
+
+### Fixed
+
+- A bug with the body text truncating on WaitingForOAuth.
+
 ## v.0.6.1
 
-- Fixes a bug with the color of the focus indicator on the favorite institution view.
+### Fixed
+
+- A bug with the color of the focus indicator on the favorite institution view.
 
 ## v.0.6.0